All the Japanese covers
Add covers for all JP games.
Covers for some "Limited edition" games are the same as the standard editions as in many cases the alt covers are an obscure shape and can't be squared easily.
Corrected some entries in the JP csv list as some titles included illegal characters that can't be used in filenames.